consider the partition_ttl when creating partitions
